Renamed KieBaseUpdater
