+++ drools-alphanetwork-compiler/src/main/java/org/drools/ancompiler/KieBaseUpdaterANC.java
@@ -1,0 +1,11 @@
+package org.drools.ancompiler;
+
+import org.drools.compiler.kie.builder.impl.KieBaseUpdater;
+
+public class KieBaseUpdaterANC implements KieBaseUpdater {
+
+    @Override
+    public void run() {
+
+    }
+}
