--- drools-alphanetwork-compiler/src/main/java/org/drools/ancompiler/UpdaterANC.java
@@ -1,6 +1,0 @@
-package org.drools.ancompiler;
-
-import org.drools.compiler.kie.builder.impl.Updater;
-
-public class UpdaterANC implements Updater {
-}
