+++ drools-alphanetwork-compiler/src/main/resources/META-INF/kie.conf
@@ -1,1 +1,1 @@
-org.drools.compiler.kie.builder.impl.UpdaterContainer = +org.drools.ancompiler.UpdaterANC
+org.drools.compiler.kie.builder.impl.KieBaseUpdaters = +org.drools.ancompiler.KieBaseUpdaterANC
