+++ drools-compiler/src/main/java/org/drools/compiler/kie/builder/impl/KieBaseUpdaters.java
@@ -1,0 +1,8 @@
+package org.drools.compiler.kie.builder.impl;
+
+import java.util.List;
+
+public interface KieBaseUpdaters {
+
+    List<KieBaseUpdater> getChildren();
+}
