--- drools-compiler/src/main/java/org/drools/compiler/kie/builder/impl/Updater.java
@@ -1,5 +1,0 @@
-package org.drools.compiler.kie.builder.impl;
-
-public interface Updater {
-
-}
