--- drools-compiler/src/main/java/org/drools/compiler/kie/builder/impl/UpdaterContainer.java
@@ -1,8 +1,0 @@
-package org.drools.compiler.kie.builder.impl;
-
-import java.util.List;
-
-public interface UpdaterContainer {
-
-    List<Updater> getChildren();
-}
