--- drools-compiler/src/main/java/org/drools/compiler/kie/builder/impl/UpdaterImpl1.java
@@ -1,5 +1,0 @@
-package org.drools.compiler.kie.builder.impl;
-
-public class UpdaterImpl1 implements Updater {
-
-}
