--- drools-compiler/src/main/java/org/drools/compiler/kie/builder/impl/UpdaterImpl2.java
@@ -1,5 +1,0 @@
-package org.drools.compiler.kie.builder.impl;
-
-public class UpdaterImpl2 implements Updater {
-
-}
