+++ drools-compiler/src/main/resources/META-INF/kie.conf
@@ -1,6 +1,4 @@
 org.kie.api.KieServices = org.drools.compiler.kie.builder.impl.KieServicesImpl
 ?org.kie.internal.builder.KnowledgeBuilderFactoryService = org.drools.compiler.builder.impl.KnowledgeBuilderFactoryServiceImpl
 ?org.kie.internal.builder.JaxbConfigurationFactoryService = org.drools.compiler.builder.impl.JaxbConfigurationFactoryServiceImpl
-org.drools.compiler.kie.builder.impl.UpdaterContainer = org.drools.compiler.kie.builder.impl.UpdaterContainerContainerImpl
-org.drools.compiler.kie.builder.impl.UpdaterContainer = +org.drools.compiler.kie.builder.impl.UpdaterImpl1
-org.drools.compiler.kie.builder.impl.UpdaterContainer = +org.drools.compiler.kie.builder.impl.UpdaterImpl2
+org.drools.compiler.kie.builder.impl.KieBaseUpdaters = org.drools.compiler.kie.builder.impl.UpdaterContainerContainerImpl
